chore: add task run context to ghost logs
